requirements.txt
flask
pyjwt
joblib
numpy
sqlalchemy